+++ src/runtime/arc.ts
@@ -23,7 +23,8 @@ import {ParticleSpec} from './particle-spec.js';
 import {StorageStub} from './storage-stub.js';
 import {Handle} from './recipe/handle.js';
 import {Particle} from './recipe/particle.js';
-import {Recipe, IsValidOptions} from './recipe/recipe.js';
+import {Recipe} from './recipe/recipe.js';
+import {IsValidOptions} from './recipe/recipe-types.js';
 import {Slot} from './recipe/slot.js';
 import {compareComparables} from './recipe/comparable.js';
 import {SlotComposer} from './slot-composer.js';
