+++ src/runtime/recipe/connection-constraint.ts
@@ -14,7 +14,8 @@ import {ParticleSpec} from '../particle-spec.js';
 import {Direction} from '../manifest-ast-nodes.js';
 import {Handle} from './handle.js';
 import {Comparable, compareArrays, compareComparables, compareStrings} from './comparable.js';
-import {Recipe, RecipeComponent, CloneMap, ToStringOptions} from './recipe.js';
+import {RecipeComponent, CloneMap, ToStringOptions} from './recipe-types.js';
+import {Recipe} from './recipe.js';
 import {Particle} from './particle.js';
 
 export abstract class EndPoint implements Comparable<EndPoint> {
