+++ src/runtime/recipe/handle-connection.ts
@@ -15,7 +15,8 @@ import {Type} from '../type.js';
 import {Handle} from './handle.js';
 import {SlotConnection} from './slot-connection.js';
 import {Particle} from './particle.js';
-import {CloneMap, IsValidOptions, Recipe, RecipeComponent, ToStringOptions, VariableMap} from './recipe.js';
+import {Recipe} from './recipe.js';
+import {CloneMap, IsValidOptions, RecipeComponent, ToStringOptions, VariableMap} from './recipe-types.js';
 import {acceptedDirections} from './recipe-util.js';
 import {TypeChecker} from './type-checker.js';
 import {compareArrays, compareComparables, compareStrings, Comparable} from './comparable.js';
