+++ src/runtime/recipe/recipe-resolver.ts
@@ -15,7 +15,8 @@ import {Handle} from './handle.js';
 import {Particle} from './particle.js';
 import {RecipeUtil} from './recipe-util.js';
 import {RecipeWalker} from './recipe-walker.js';
-import {Recipe, IsValidOptions} from './recipe.js';
+import {Recipe} from './recipe.js';
+import {IsValidOptions} from './recipe-types.js';
 import {ConnectionConstraint, InstanceEndPoint} from './connection-constraint.js';
 import {SlotConnection} from './slot-connection.js';
 import {SlotUtils} from './slot-utils.js';
