+++ src/runtime/recipe/recipe.ts
@@ -28,14 +28,15 @@ import {compareComparables} from './comparable.js';
 import {Cloneable} from './walker.js';
 import {Dictionary} from '../hot.js';
 import {RecipeInterface} from './recipe-interface.js';
+import {IsValidOptions, RecipeComponent, CloneMap, ToStringOptions} from './recipe-types.js';
 
-export type RecipeComponent = Particle | Handle | HandleConnection | Slot | SlotConnection | EndPoint;
-export type CloneMap = Map<RecipeComponent, RecipeComponent>;
-export type VariableMap = Map<TypeVariableInfo|Schema, TypeVariableInfo|Schema>;
+// export type RecipeComponent = Particle | Handle | HandleConnection | Slot | SlotConnection | EndPoint;
+// export type CloneMap = Map<RecipeComponent, RecipeComponent>;
+// export type VariableMap = Map<TypeVariableInfo|Schema, TypeVariableInfo|Schema>;
 
-export type IsResolvedOptions = {showUnresolved?: boolean, details?: string[]}; // TODO(lindner): standardize details
-export type IsValidOptions = {errors?: Map<Recipe | RecipeComponent, string>};
-export type ToStringOptions = {showUnresolved?: boolean, hideFields?: boolean, details?: string[]};
+// export type IsResolvedOptions = {showUnresolved?: boolean, details?: string[]}; // TODO(lindner): standardize details
+// export type IsValidOptions = {errors?: Map<Recipe | RecipeComponent, string>};
+// export type ToStringOptions = {showUnresolved?: boolean, hideFields?: boolean, details?: string[]};
 
 export class Recipe implements Cloneable<Recipe>, RecipeInterface {
   private readonly _requires: RequireSection[] = [];
